% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountcodepm.R
\name{rds_accountcode_pm}
\alias{rds_accountcode_pm}
\title{付款单科目编码更新}
\usage{
rds_accountcode_pm(token = "5148FAAF-E45C-43D3-9F96-9F943F4F341B")
}
\arguments{
\item{token}{访问口令}
}
\value{
无返回值
}
\description{
付款单科目编码更新
}
