% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accountcodepo.R
\name{rds_accountcode_po}
\alias{rds_accountcode_po}
\title{更新采购订单付款计划科目编码}
\usage{
rds_accountcode_po(token = "5148FAAF-E45C-43D3-9F96-9F943F4F341B")
}
\arguments{
\item{token}{访问口令}
}
\value{
无返回值
}
\description{
更新采购订单付款计划科目编码
}
